pyproject.toml
[project]
name = "mapping"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "django-qrcode>=0.3",
    "pillow>=11.3.0",
    "qrcode[pil]>=8.2",
    "tomli>=2.3.0",
    "tomli-w>=1.2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-mock>=3.15.1",
    "pytidylib>=0.3.2",
]
